Split out a helper object for article metadata
